[FIX] move_to_s3 method
Fixed the method to move all attachments to S3. It can be called from
xmlrpc or S3. Since we force the storage to be file before moving, the
permissions are checked by the upstream method. To delete the files
localy, the user must pass the argument delete_fs as True